fix(migtd/migration): tighten MigtdMigrationInformation parser check
The policy_v2 length check accepted a full-size payload with
has_init_data == 0 — a band-aid for a test whose fixture started
writing the full struct after the buffer-unification refactor grew
its size. Replace it with a symmetric check that rejects whenever the
flag and length disagree:

    (has_init_data == 1) != (data_length == full_size)  =>  reject

Move the 8 parser tests from rebinding.rs (only built under
main+policy_v2+vmcall-raw, never run by `xtask lib-test`) to a
policy_v2-gated module in mod.rs. The shared build_mig_info fixture
is now layout-aware so the same tests cover both vmcall-raw (56/568)
and non-vmcall-raw (72/584) struct layouts.

Make MIGTD_MIGRATION_INFO_HEADER_SIZE pub(crate) and update two
fixtures to emit short-form payloads that satisfy the new check:
data::create_mig_info_hob (policy_v2 only) and
session::build_migration_payload. The production HOB consumer
create_migration_information stays layout-neutral; consistency is now
enforced inside the parser.
