+++ src/migtd/src/migration/rebinding.rs
@@ -522,123 +522,3 @@ async fn tls_session_read_exact(
     }
     Ok(())
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::migration::{td_info_mrowner, td_info_mrownerconfig};
-    use alloc::vec;
-
-    /// Create a 512-byte TDINFO_STRUCT with known mrowner and mrownerconfig.
-    fn make_tdinfo(mrowner: &[u8; 48], mrownerconfig: &[u8; 48]) -> [u8; TD_INFO_SIZE] {
-        let mut tdinfo = [0u8; TD_INFO_SIZE];
-        // mrowner at offset 112..160
-        tdinfo[112..160].copy_from_slice(mrowner);
-        // mrownerconfig at offset 160..208
-        tdinfo[160..208].copy_from_slice(mrownerconfig);
-        tdinfo
-    }
-
-    /// Build a MigtdMigrationInformation byte buffer. When `init_tdinfo` is `Some`, the
-    /// raw TDINFO bytes are appended directly (no envelope, per the new wire
-    /// contract).
-    fn build_mig_info(
-        mig_request_id: u64,
-        migration_source: u8,
-        has_init_data: u8,
-        uuid: [u64; 4],
-        binding_handle: u64,
-        init_tdinfo: Option<&[u8]>,
-    ) -> Vec<u8> {
-        let mut buf = Vec::new();
-        buf.extend_from_slice(&mig_request_id.to_le_bytes()); // 0..8
-        buf.push(migration_source); // 8
-        buf.push(has_init_data); // 9
-        buf.extend_from_slice(&[0u8; 6]); // 10..16 reserved
-        for u in &uuid {
-            buf.extend_from_slice(&u.to_le_bytes()); // 16..48
-        }
-        buf.extend_from_slice(&binding_handle.to_le_bytes()); // 48..56
-        if let Some(data) = init_tdinfo {
-            buf.extend_from_slice(data); // 56..56+512
-        }
-        buf
-    }
-
-    #[test]
-    fn test_mig_info_no_init_data() {
-        let buf = build_mig_info(42, 1, 0, [1, 2, 3, 4], 99, None);
-        let info = MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf)
-            .expect("should parse");
-        assert_eq!(info.mig_request_id, 42);
-        assert_eq!(info.migration_source, 1);
-        assert_eq!(info.has_init_data, 0);
-        assert_eq!(info.target_td_uuid, [1, 2, 3, 4]);
-        assert_eq!(info.binding_handle, 99);
-        assert!(info.init_td_info_if_present().is_none());
-        // Padded tail is zero.
-        assert_eq!(info.init_td_info, [0u8; TD_INFO_SIZE]);
-    }
-
-    #[test]
-    fn test_mig_info_with_init_data() {
-        let tdinfo = make_tdinfo(&[0xCAu8; 48], &[0xFEu8; 48]);
-        let buf = build_mig_info(7, 0, 1, [10, 20, 30, 40], 55, Some(&tdinfo));
-        let info = MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf)
-            .expect("should parse with init data");
-        assert_eq!(info.mig_request_id, 7);
-        assert_eq!(info.has_init_data, 1);
-        let init = info.init_td_info_if_present().expect("should be present");
-        assert_eq!(init, &tdinfo);
-        assert_eq!(td_info_mrowner(init), &[0xCAu8; 48]);
-        assert_eq!(td_info_mrownerconfig(init), &[0xFEu8; 48]);
-    }
-
-    #[test]
-    fn test_mig_info_rejects_short_buffer() {
-        assert!(MigtdMigrationInformation::read_from_bytes(10, &[0u8; 10]).is_err());
-        assert!(MigtdMigrationInformation::read_from_bytes(55, &[0u8; 55]).is_err());
-        // 55 < 56
-    }
-
-    #[test]
-    fn test_mig_info_rejects_nonzero_reserved() {
-        let mut buf = build_mig_info(1, 0, 0, [0; 4], 0, None);
-        buf[10] = 0xFF; // reserved byte not zero
-        assert!(MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf).is_err());
-    }
-
-    #[test]
-    fn test_mig_info_rejects_has_init_data_without_tail() {
-        // has_init_data=1 but no tail bytes following → flag/length mismatch
-        let buf = build_mig_info(1, 0, 1, [0; 4], 0, None);
-        assert!(MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf).is_err());
-    }
-
-    #[test]
-    fn test_mig_info_rejects_full_form_without_flag() {
-        // has_init_data=0 but full 568-byte buffer → flag/length mismatch
-        let tdinfo = [0u8; TD_INFO_SIZE];
-        let buf = build_mig_info(1, 0, 0, [0; 4], 0, Some(&tdinfo));
-        assert!(MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf).is_err());
-    }
-
-    #[test]
-    fn test_mig_info_rejects_invalid_has_init_data() {
-        // has_init_data must be 0 or 1
-        let mut buf = build_mig_info(1, 0, 0, [0; 4], 0, None);
-        buf[9] = 2;
-        assert!(MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf).is_err());
-        buf[9] = 0xFF;
-        assert!(MigtdMigrationInformation::read_from_bytes(buf.len() as u32, &buf).is_err());
-    }
-
-    #[test]
-    fn test_mig_info_rejects_unexpected_length() {
-        // Only short (56) or full (568) lengths are accepted.
-        let mid_buf = vec![0u8; 100];
-        assert!(
-            MigtdMigrationInformation::read_from_bytes(mid_buf.len() as u32, &mid_buf).is_err()
-        );
-    }
-}
